refactor the codec package
