+++ ratatouille-common/src/main/scala/fr/xebia/ldi/ratatouille/model/Dinner.scala
@@ -1,6 +1,6 @@
-package fr.xebia.ldi.ratatouille.codec
+package fr.xebia.ldi.ratatouille.model
 
-import fr.xebia.ldi.ratatouille.codec.Dinner.{Client, Command, Moment}
+import fr.xebia.ldi.ratatouille.model.Dinner.{Client, Command, Moment}
 
 /**
   * Created by loicmdivad.
